+++ apps/web/shared/types/user.ts
@@ -1,11 +1,5 @@
 import type { GalgameCard } from './galgame'
 
-export interface UserTopic {
-  id: number
-  title: string
-  created: Date | string
-}
-
 export type UserGalgame = GalgameCard
 
 export interface UserGalgameResource {
@@ -21,19 +15,5 @@ export interface UserGalgameResource {
   password: string
   note: string
   status: number
-  created: Date | string
-}
-
-export interface UserReply {
-  topic_id: number
-  floor: number
-  content: string
-  created: Date | string
-}
-
-export interface UserComment {
-  id: number
-  topic_id: number
-  content: string
   created: Date | string
 }